pydantic-ai-quickstart/requirements.txt
pydantic-ai>=0.0.49
python-dotenv>=1.0.0
httpx>=0.25.0
pydantic>=2.5.0